Re-do Attr according to modern standards
This completely replaces the Attr implementation, and all related code. It is now based on the latest spec, in particular no longer inheriting from Node. It uses webidl2js to generate a lot of the boilerplate.

A couple other methods were updated with better IDL compliance in the process of getting tests to pass:
- Node.prototype.insertBefore
- Node.prototype.replaceChild
- Node.prototype.appendChild

A number of tests were removed since they test for things that are no longer true, like Attr having child nodes. Some of these that have changed results were introduced as new to-upstream web-platform-tests.